Refactor code to use separate context for pool name
